Fixed redux bug. Setup more constant files for UI values, redux action types, and type interfaces for TypeScript. Fixed typo in Readme.
