+++ src/actions/constants.tsx
@@ -1,0 +1,2 @@
+export const CHANGE_FILTERS = 'CHANGE_FILTERS';
+export const SUBMIT_SEARCH = 'SUBMIT_SEARCH';
