+++ src/components/UIConstants.tsx
@@ -1,0 +1,3 @@
+export const MIN_COURSE_CODE = 0;
+export const MAX_COURSE_CODE = 800;
+export const SLIDER_STEP_SIZE = 100;
