+++ src/reducers/Browse.tsx
@@ -1,4 +1,5 @@
-import { Filters } from '../components/Browse'; // eslint-disable-line no-unused-vars
+import { CHANGE_FILTERS, SUBMIT_SEARCH } from '../actions/constants';
+import { Filters } from '../type-interfaces/Browse'; // eslint-disable-line no-unused-vars
 
 const searchFilteringDefault: Filters = {
   query: 'cpsc',
@@ -7,7 +8,7 @@ const searchFilteringDefault: Filters = {
 };
 export const searchFiltering = (store: Filters = searchFilteringDefault, action: any) => {
   switch (action.type) {
-    case 'CHANGE_SEARCH':
+    case CHANGE_FILTERS:
       return {
         ...store,
         query: action.query,
@@ -23,7 +24,7 @@ export const searchFiltering = (store: Filters = searchFilteringDefault, action:
 const searchResultsDefault: { courses: object[] } = { courses: [] };
 export const searchRetrieval = (store: { courses: object[] } = searchResultsDefault, action: any) => {
   switch (action.type) {
-    case 'SUBMIT_SEARCH':
+    case SUBMIT_SEARCH:
       return store; // TODO add actual behavious of searching
     default:
       return store;
